sigh ...
Because I tried to create the project initially as 1Q instead of OQ,
some weird things were happening. So, I re-created project and moved
files over. This of course caused me to lose my commit history, but
it’s just a test project so not going to worry about it for now.